Removing unused files
