+++ gpu4pyscf/pbc/df/int3c2e.py
@@ -1,4 +1,4 @@
-# Copyright 2024-2025 The PySCF Developers. All Rights Reserved.
+# Copyright 2024-2026 The PySCF Developers. All Rights Reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -28,7 +28,8 @@ from pyscf.pbc.lib.kpts_helper import is_zero
 from pyscf.pbc.df.rsdf_builder import estimate_ke_cutoff_for_omega
 from gpu4pyscf.pbc.tools.k2gamma import kpts_to_kmesh
 from gpu4pyscf.lib import logger
-from gpu4pyscf.lib.cupy_helper import contract, asarray, transpose_sum, ndarray
+from gpu4pyscf.lib.cupy_helper import (
+    contract, asarray, transpose_sum, ndarray, empty_aligned)
 from gpu4pyscf.lib.utils import splits_by_blocksize
 from gpu4pyscf.gto.mole import (
     groupby, PTR_BAS_COORD, extract_pgto_params, SortedCell,
@@ -53,7 +54,9 @@ libpbc.PBCcontract_int3c2e_auxvec.restype = ctypes.c_int
 LMAX = 4
 L_AUX_MAX = 6
 THREADS = 256
-POOL_SIZE = 262144
+POOL_SIZE = 16384
+MAX_IMGS_PER_TASK = 31
+GOUT_WIDTH = 54
 
 def sr_aux_e2(cell, auxcell, omega, kpts=None, bvk_kmesh=None, j_only=False):
     r'''
@@ -84,22 +87,19 @@ def sr_aux_e2(cell, auxcell, omega, kpts=None, bvk_kmesh=None, j_only=False):
     auxcell = int3c2e_opt.auxcell
     bvk_ncells = len(int3c2e_opt.bvkmesh_Ls)
 
-    eval_j3c, aux_sorting = int3c2e_opt.int3c2e_evaluator()[:2]
+    eval_j3c = int3c2e_opt.int3c2e_evaluator()[0]
     pair_address = int3c2e_opt.pair_and_diag_indices()[0]
     aux_coeff = auxcell.ctr_coeff
-    aux_coeff, tmp = cp.empty_like(aux_coeff), aux_coeff
-    aux_coeff[aux_sorting] = tmp
-    tmp = None
     j3c = eval_j3c()
 
     if is_gamma_point:
-        j3c = j3c[:,:,0].dot(aux_coeff)
+        j3c = j3c[:,0,:].dot(aux_coeff)
         out = cp.zeros((nao, nao, naux))
         i, j = divmod(pair_address, nao*bvk_ncells)
         out[j, i] = out[i, j] = j3c
 
     elif j_only:
-        j3c = j3c.sum(axis=2).dot(aux_coeff)
+        j3c = j3c.sum(axis=1).dot(aux_coeff)
         bvkmesh_Ls = cp.asarray(int3c2e_opt.bvkmesh_Ls)
         kpts = cp.asarray(kpts).reshape(-1, 3)
         expLk = cp.exp(1j*bvkmesh_Ls.dot(kpts.T))
@@ -110,7 +110,7 @@ def sr_aux_e2(cell, auxcell, omega, kpts=None, bvk_kmesh=None, j_only=False):
         out = out.transpose(0,2,3,1)
 
     else:
-        j3c = contract('tpL,pq->tqL', j3c, aux_coeff)
+        j3c = contract('tLp,pq->tqL', j3c, aux_coeff)
         bvkmesh_Ls = cp.asarray(int3c2e_opt.bvkmesh_Ls)
         kpts = cp.asarray(kpts).reshape(-1, 3)
         expLk = cp.exp(1j*bvkmesh_Ls.dot(kpts.T))
@@ -201,7 +201,7 @@ def fill_triu_bvk(a, nao, bvk_kmesh, pair_address=None, conj_mapping=None, bvk_a
 class SRInt3c2eOpt:
     def __init__(self, cell, auxcell, omega, bvk_kmesh=None):
         omega = abs(omega)
-        self.omega = -omega
+        self.omega = omega
         self.cell = SortedCell.from_cell(
             cell, allow_replica=True, allow_split_seg_contraction=False)
         assert self.cell.uniq_l_ctr[:,0].max() <= LMAX
@@ -248,8 +248,8 @@ class SRInt3c2eOpt:
         self.bvk_auxcell = bvk_auxcell
 
         if self.rcut is None:
-            omega = -abs(self.omega)
-            rcut = max(estimate_rcut(cell, auxcell, omega).max(), cell.rcut, auxcell.rcut)
+            omega = -self.omega
+            rcut = max(estimate_rcut(cell, auxcell, omega).max(), cell.rcut)
             self.rcut = rcut
         Ls = asarray(bvkcell.get_lattice_Ls(rcut=self.rcut))
         Ls = Ls[cp.linalg.norm(Ls-.5, axis=1).argsort()]
@@ -263,7 +263,7 @@ class SRInt3c2eOpt:
         off = _bas[bvkcell.nbas,PTR_EXP] - bvk_auxcell._bas[0,PTR_EXP]
         _bas[bvkcell.nbas:,PTR_BAS_COORD] += off
         ao_loc = bvkcell.ao_loc
-        aux_loc = bvk_auxcell.ao_loc
+        aux_loc = bvk_auxcell.ao_loc_nr(cart=True)
         ao_loc = cp.asarray(_conc_locs(ao_loc, aux_loc), dtype=np.int32)
         self._int3c2e_envs = PBCIntEnvVars.new(
             cell.natm, cell.nbas, bvk_ncells, nimgs, _atm, _bas, _env, ao_loc, Ls)
@@ -340,7 +340,7 @@ class SRInt3c2eOpt:
         if self._mesh is not None:
             return self._mesh
         cell = self.cell
-        omega = abs(self.omega)
+        omega = self.omega
         ke_cutoff = estimate_ke_cutoff_for_omega(cell, omega)
         mesh = cell.cutoff_to_mesh(ke_cutoff)
         mesh = cell.symmetrize_mesh(mesh)
@@ -379,15 +379,15 @@ class SRInt3c2eOpt:
         auxcell = self.auxcell
         bvk_ncells = np.prod(self.bvk_kmesh)
 
-        nsp_per_block, gout_stride, shm_size = int3c2e_scheme(gout_width=54)
+        nsp_per_block, gout_stride, shm_size = int3c2e_scheme(gout_width=GOUT_WIDTH)
         lmax = cell.uniq_l_ctr[:,0].max()
         laux = auxcell.uniq_l_ctr[:,0].max()
         shm_size_max = shm_size[:laux+1,:lmax+1,:lmax+1].max()
         if bas_ij_aggregated is None:
-            bas_ij_idx, shl_pair_offsets = cell.aggregate_shl_pairs(
-                self.bas_ij_cache, nsp_per_block[0])
+            bas_ij_idx, batched_shl_pair_offsets = cell.aggregate_shl_pairs(
+                self.bas_ij_cache, 1000000)
         else:
-            bas_ij_idx, shl_pair_offsets = bas_ij_aggregated
+            bas_ij_idx, batched_shl_pair_offsets = bas_ij_aggregated
 
         # For each primitive shell-pair in bas_ij_idx, ao_pair_loc points to the
         # addresses of first element for the contracted pair-GTOs. In each
@@ -397,64 +397,56 @@ class SRInt3c2eOpt:
             cart = cell.cell.cart
         ao_pair_loc = get_ao_pair_loc(cell.uniq_l_ctr[:,0], self.bas_ij_cache, cart)
 
-        if ao_pair_batch_size is None:
-            pair_splits = [0, len(shl_pair_offsets)-1]
-            ao_pair_offsets = [0, ao_pair_loc[-1].get()]
-        else:
-            ao_pair_offsets = ao_pair_loc[shl_pair_offsets].get()
-            pair_splits = splits_by_blocksize(ao_pair_offsets, ao_pair_batch_size)
-            ao_pair_offsets = ao_pair_offsets[pair_splits]
-        shl_pair_offsets = cp.asnumpy(shl_pair_offsets)
-
-        # Split auxbasis in the unit cell than the bvk-cell
-        aux_loc = auxcell.ao_loc
         uniq_l_ctr_aux = auxcell.uniq_l_ctr
         l_ctr_aux_offsets = np.append(0, np.cumsum(auxcell.l_ctr_counts))
+        # Split auxbasis in the unit cell
         if aux_batch_size is None:
-            ksh_offsets_cpu = l_ctr_aux_offsets
-            aux_splits = [0, len(ksh_offsets_cpu)-1]
+            _aux_batch_size = POOL_SIZE // bvk_ncells
         else:
-            l_ctr_aux_offsets, uniq_l_ctr_aux = _split_l_ctr_pattern(
-                l_ctr_aux_offsets, uniq_l_ctr_aux, aux_batch_size)
-            ksh_offsets_cpu = l_ctr_aux_offsets
-            aux_splits = range(len(ksh_offsets_cpu))
-        aux_offsets = aux_loc[ksh_offsets_cpu[aux_splits]]
-        aux_sorting = argsort_aux(l_ctr_aux_offsets, uniq_l_ctr_aux)
+            _aux_batch_size = aux_batch_size
+        l_ctr_aux_offsets, uniq_l_ctr_aux = _split_l_ctr_pattern(
+            l_ctr_aux_offsets, uniq_l_ctr_aux, _aux_batch_size)
 
-        ksh_idx = _aggregate_bas_idx(
-            l_ctr_aux_offsets, uniq_l_ctr_aux, bvk_ncells, auxcell.nbas)[1]
-        ksh_idx += self.bvkcell.nbas
-        ksh_offsets_gpu = cp.asarray(ksh_offsets_cpu, dtype=np.int32)
-        shl_pair_batches = len(ao_pair_offsets) - 1
-        aux_batches = len(aux_offsets) - 1
+        aux_loc = auxcell.ao_loc
+        aux_groups, aux_offsets = _group_ksh_batches(
+            l_ctr_aux_offsets, uniq_l_ctr_aux, aux_loc, aux_batch_size)
+        aux_sorting = slice(None)
+
+        ksh_dims = l_ctr_aux_offsets[1:] - l_ctr_aux_offsets[:-1]
+        pair_per_block = POOL_SIZE // (ksh_dims.max() * bvk_ncells)
+        assert pair_per_block > 0, 'aux_batch_size is too large'
+        shl_pair_batches = len(batched_shl_pair_offsets) - 1
+        pair_per_block = min(pair_per_block, max(4, shl_pair_batches//20))
+        shl_pair_groups, ao_pair_offsets = _group_shl_pair_batches(
+            batched_shl_pair_offsets, ao_pair_loc, ao_pair_batch_size, pair_per_block)
+
         logger.debug1(self.cell, 'sp_batches = %d, ksh_batches = %d',
-                      shl_pair_batches, aux_batches)
+                      len(shl_pair_groups), len(aux_groups))
         diffuse_exps = cp.asarray(self.diffuse_exps)
         diffuse_coefs = cp.asarray(self.diffuse_coefs)
-        atom_aux_exps = cp.asarray(diffuse_exps_by_atom(auxcell), dtype=np.float32)
         log_cutoff = math.log(self.cutoff)
 
         workers = gpu_specs['multiProcessorCount']
-        pool = cp.empty((workers, POOL_SIZE), dtype=np.uint32)
+        pool = cp.empty(workers * POOL_SIZE*(MAX_IMGS_PER_TASK+2) + 1, dtype=np.uint32)
+        head = pool[-1:]
+        task_pool = empty_aligned((workers, POOL_SIZE*16), np.int32, alignment=128)
+        c2s_pool = cp.empty((workers, THREADS*GOUT_WIDTH))
         int3c2e_envs = self.int3c2e_envs
         img_idx = cp.asarray(self.img_idx)
         img_offsets = cp.asarray(self.img_offsets)
         kern = libpbc.PBCsr_int3c2e_latsum23
 
         def evaluate_j3c(shl_pair_batch_id=0, aux_batch_id=0, out=None):
-            pair_split0 = pair_splits[shl_pair_batch_id]
-            pair_split1 = pair_splits[shl_pair_batch_id+1]
-            shl_pair0 = shl_pair_offsets[pair_split0]
-            shl_pair1 = shl_pair_offsets[pair_split1]
-            ao_pair_offset = ao_pair_offsets[shl_pair_batch_id]
-            nao_pair = ao_pair_offsets[shl_pair_batch_id+1] - ao_pair_offset
+            shl_pair_group = shl_pair_groups[shl_pair_batch_id]
+            shl_pair0, shl_pair1 = shl_pair_group.shl_range
+            ao_pair_offset = shl_pair_group.ao_range[0]
+            nao_pairs = shl_pair_group.ao_range[1] - ao_pair_offset
 
             # Indexing the aux-basis within the first cell
-            aux_split0 = aux_splits[aux_batch_id]
-            aux_split1 = aux_splits[aux_batch_id+1]
-            aux_ao_offset = aux_offsets[aux_batch_id]
-            naux = aux_offsets[aux_batch_id+1] - aux_ao_offset
-            out = ndarray((nao_pair, naux, bvk_ncells), buffer=out)
+            aux_group = aux_groups[aux_batch_id]
+            aux_ao_offset = aux_group.ao_range[0]
+            naux = aux_group.ao_range[1] - aux_ao_offset
+            out = ndarray((nao_pairs, bvk_ncells, naux), buffer=out)
             # The output buffer must be initialized because integral screening
             # based on SR integrals is performed in the kernel, and certain ~0
             # shell-tritets are not evaluated, leaving the output buffer untouched
@@ -465,23 +457,25 @@ class SRInt3c2eOpt:
                 ctypes.cast(out.data.ptr, ctypes.c_void_p),
                 ctypes.byref(int3c2e_envs),
                 ctypes.cast(pool.data.ptr, ctypes.c_void_p),
+                ctypes.cast(task_pool.data.ptr, ctypes.c_void_p),
+                ctypes.cast(c2s_pool.data.ptr, ctypes.c_void_p),
+                ctypes.cast(head.data.ptr, ctypes.c_void_p),
                 ctypes.c_int(shm_size_max),
-                ctypes.c_int(shl_pair1 - shl_pair0),
-                ctypes.c_int(aux_split1 - aux_split0),
-                ctypes.cast(bas_ij_idx[shl_pair0:].data.ptr, ctypes.c_void_p),
-                ctypes.cast(ksh_offsets_gpu[aux_split0:].data.ptr, ctypes.c_void_p),
-                ctypes.cast(ksh_idx.data.ptr, ctypes.c_void_p),
+                ctypes.c_int(len(shl_pair_group.sub_batch_offsets) - 1),
+                ctypes.c_int(len(aux_group.sub_batch_offsets) - 1),
+                ctypes.cast(bas_ij_idx.data.ptr, ctypes.c_void_p),
+                ctypes.cast(shl_pair_group.sub_batch_offsets.data.ptr, ctypes.c_void_p),
+                ctypes.cast(aux_group.sub_batch_offsets.data.ptr, ctypes.c_void_p),
                 ctypes.cast(img_idx.data.ptr, ctypes.c_void_p),
-                ctypes.cast(img_offsets[shl_pair0:].data.ptr, ctypes.c_void_p),
+                ctypes.cast(img_offsets.data.ptr, ctypes.c_void_p),
                 ctypes.cast(gout_stride.data.ptr, ctypes.c_void_p),
-                ctypes.cast(ao_pair_loc[shl_pair0:].data.ptr, ctypes.c_void_p),
-                ctypes.c_int(ao_pair_offset),
+                ctypes.cast(ao_pair_loc.data.ptr, ctypes.c_void_p),
+                ctypes.c_int(ao_pair_offset), ctypes.c_int(nao_pairs),
                 ctypes.c_int(aux_ao_offset),
-                ctypes.c_int(naux * bvk_ncells),
+                ctypes.c_int(auxcell.nbas), ctypes.c_int(naux),
                 ctypes.c_int(not cart),
                 ctypes.cast(diffuse_exps.data.ptr, ctypes.c_void_p),
                 ctypes.cast(diffuse_coefs.data.ptr, ctypes.c_void_p),
-                ctypes.cast(atom_aux_exps.data.ptr, ctypes.c_void_p),
                 ctypes.c_float(log_cutoff))
             if err != 0:
                 raise RuntimeError('fill_int3c2e kernel')
@@ -511,59 +505,47 @@ class SRInt3c2eOpt:
 
         cell = self.cell
         auxcell = self.auxcell
-        bvk_ncells = len(self.bvkmesh_Ls)
 
         nsp_per_block, gout_stride, shm_size = int3c2e_scheme()
         lmax = cell.uniq_l_ctr[:,0].max()
         laux = auxcell.uniq_l_ctr[:,0].max()
         shm_size_max = shm_size[:laux+1,:lmax+1,:lmax+1].max()
-        bas_ij_idx, shl_pair_offsets = cell.aggregate_shl_pairs(self.bas_ij_cache, 256)
-
-        l_ctr_aux_offsets = np.append(0, np.cumsum(auxcell.l_ctr_counts))
-        l_ctr_aux_offsets, uniq_l_ctr_aux = _split_l_ctr_pattern(
-            l_ctr_aux_offsets, auxcell.uniq_l_ctr, 32)
-        aux_sorting = argsort_aux(l_ctr_aux_offsets, uniq_l_ctr_aux)
-
-        ksh_idx = _aggregate_bas_idx(
-            l_ctr_aux_offsets, uniq_l_ctr_aux, bvk_ncells, auxcell.nbas)[1]
-        ksh_idx += self.bvkcell.nbas
-        ksh_offsets = cp.asarray(l_ctr_aux_offsets, dtype=np.int32)
-        shl_pair_batches = len(shl_pair_offsets) - 1
-        aux_batches = len(ksh_offsets) - 1
+        bas_ij_idx, shl_pair_offsets = cell.aggregate_shl_pairs(self.bas_ij_cache, 1000000)
 
         diffuse_exps = cp.asarray(self.diffuse_exps)
         diffuse_coefs = cp.asarray(self.diffuse_coefs)
-        atom_aux_exps = cp.asarray(diffuse_exps_by_atom(auxcell), dtype=np.float32)
         log_cutoff = math.log(self.cutoff)
 
         workers = gpu_specs['multiProcessorCount']
-        pool = cp.empty((workers, POOL_SIZE), dtype=np.uint32)
+        pool = cp.empty(workers * POOL_SIZE*(MAX_IMGS_PER_TASK+2) + 1, dtype=np.uint32)
+        head = pool[-1:]
+        task_pool = empty_aligned((workers, POOL_SIZE*16), np.int32, alignment=128)
         int3c2e_envs = self.int3c2e_envs
+        img_idx = cp.asarray(self.img_idx)
+        img_offsets = cp.asarray(self.img_offsets)
 
         naux = auxcell.nao
-        vj_aux = cp.zeros((naux, bvk_ncells))
+        vj_aux = cp.zeros(naux)
         err = libpbc.PBCcontract_int3c2e_dm(
             ctypes.cast(vj_aux.data.ptr, ctypes.c_void_p),
             ctypes.cast(dm.data.ptr, ctypes.c_void_p),
             ctypes.byref(int3c2e_envs),
             ctypes.cast(pool.data.ptr, ctypes.c_void_p),
+            ctypes.cast(task_pool.data.ptr, ctypes.c_void_p),
+            ctypes.cast(head.data.ptr, ctypes.c_void_p),
             ctypes.c_int(shm_size_max),
-            ctypes.c_int(shl_pair_batches),
-            ctypes.c_int(aux_batches),
-            ctypes.cast(shl_pair_offsets.data.ptr, ctypes.c_void_p),
+            ctypes.c_int(len(shl_pair_offsets) - 1),
+            ctypes.c_int(auxcell.nbas),
             ctypes.cast(bas_ij_idx.data.ptr, ctypes.c_void_p),
-            ctypes.cast(ksh_offsets.data.ptr, ctypes.c_void_p),
-            ctypes.cast(ksh_idx.data.ptr, ctypes.c_void_p),
-            ctypes.cast(self.img_idx.data.ptr, ctypes.c_void_p),
-            ctypes.cast(self.img_offsets.data.ptr, ctypes.c_void_p),
+            ctypes.cast(shl_pair_offsets.data.ptr, ctypes.c_void_p),
+            ctypes.cast(img_idx.data.ptr, ctypes.c_void_p),
+            ctypes.cast(img_offsets.data.ptr, ctypes.c_void_p),
             ctypes.cast(gout_stride.data.ptr, ctypes.c_void_p),
             ctypes.cast(diffuse_exps.data.ptr, ctypes.c_void_p),
             ctypes.cast(diffuse_coefs.data.ptr, ctypes.c_void_p),
-            ctypes.cast(atom_aux_exps.data.ptr, ctypes.c_void_p),
             ctypes.c_float(log_cutoff))
         if err != 0:
             raise RuntimeError('contract_int3c2e_dm failed')
-        vj_aux = vj_aux.sum(axis=1)[aux_sorting]
         if hermi == 1:
             vj_aux *= 2
         if sort_output:
@@ -589,20 +571,19 @@ class SRInt3c2eOpt:
         bas_ij_idx = cell.aggregate_shl_pairs(self.bas_ij_cache, 1000000)[0]
 
         l_ctr_aux_offsets = np.append(0, np.cumsum(auxcell.l_ctr_counts))
-        ksh_idx = _aggregate_bas_idx(
-            l_ctr_aux_offsets, auxcell.uniq_l_ctr, bvk_ncells, auxcell.nbas)[1]
-        ksh_idx += self.bvkcell.nbas
         ksh_offsets = cp.asarray(l_ctr_aux_offsets, dtype=np.int32)
-        aux_batches = len(ksh_offsets) - 1
 
         diffuse_exps = cp.asarray(self.diffuse_exps)
         diffuse_coefs = cp.asarray(self.diffuse_coefs)
-        atom_aux_exps = cp.asarray(diffuse_exps_by_atom(auxcell), dtype=np.float32)
         log_cutoff = math.log(self.cutoff)
 
         workers = gpu_specs['multiProcessorCount']
-        pool = cp.empty((workers, POOL_SIZE), dtype=np.uint32)
+        pool = cp.empty(workers * POOL_SIZE*(MAX_IMGS_PER_TASK+2) + 1, dtype=np.uint32)
+        head = pool[-1:]
+        task_pool = empty_aligned((workers, POOL_SIZE*16), np.int32, alignment=128)
         int3c2e_envs = self.int3c2e_envs
+        img_idx = cp.asarray(self.img_idx)
+        img_offsets = cp.asarray(self.img_offsets)
 
         nao = cell.nao
         vj = cp.zeros((nao, bvk_ncells, nao))
@@ -611,18 +592,19 @@ class SRInt3c2eOpt:
             ctypes.cast(auxvec.data.ptr, ctypes.c_void_p),
             ctypes.byref(int3c2e_envs),
             ctypes.cast(pool.data.ptr, ctypes.c_void_p),
+            ctypes.cast(task_pool.data.ptr, ctypes.c_void_p),
+            ctypes.cast(head.data.ptr, ctypes.c_void_p),
             ctypes.c_int(shm_size_max),
             ctypes.c_int(len(bas_ij_idx)),
-            ctypes.c_int(aux_batches),
+            ctypes.c_int(len(ksh_offsets) - 1),
+            ctypes.c_int(auxcell.nbas),
             ctypes.cast(bas_ij_idx.data.ptr, ctypes.c_void_p),
             ctypes.cast(ksh_offsets.data.ptr, ctypes.c_void_p),
-            ctypes.cast(ksh_idx.data.ptr, ctypes.c_void_p),
-            ctypes.cast(self.img_idx.data.ptr, ctypes.c_void_p),
-            ctypes.cast(self.img_offsets.data.ptr, ctypes.c_void_p),
+            ctypes.cast(img_idx.data.ptr, ctypes.c_void_p),
+            ctypes.cast(img_offsets.data.ptr, ctypes.c_void_p),
             ctypes.cast(gout_stride.data.ptr, ctypes.c_void_p),
             ctypes.cast(diffuse_exps.data.ptr, ctypes.c_void_p),
             ctypes.cast(diffuse_coefs.data.ptr, ctypes.c_void_p),
-            ctypes.cast(atom_aux_exps.data.ptr, ctypes.c_void_p),
             ctypes.c_float(log_cutoff))
         if err != 0:
             raise RuntimeError('contract_int3c2e_auxvec failed')
@@ -720,14 +702,6 @@ def estimate_rcut(cell, auxcell, omega):
     rcut = r0
     return rcut
 
-def diffuse_exps_by_atom(cell):
-    '''Find the most diffuse functions on each atom'''
-    exps, cs = extract_pgto_params(cell, 'diffuse')
-    ls = cell._bas[:,ANG_OF]
-    r2 = np.log(cs**2 / cell.precision * 10**ls + 1e-200) / exps
-    idx = groupby(cell._bas[:,ATOM_OF], r2, 'argmax')
-    return exps[idx]
-
 def _aggregate_shl_pairs(img_idx_cache, nsp_per_block):
     sp_img_idx = []
     sp_img_offsets = []
@@ -753,20 +727,86 @@ def _aggregate_shl_pairs(img_idx_cache, nsp_per_block):
     shl_pair_offsets = cp.asarray(cp.hstack(shl_pair_offsets), dtype=np.int32)
     return shl_pair_offsets, bas_ij_idx, sp_img_idx, sp_img_offsets
 
-def _aggregate_bas_idx(l_ctr_offsets, uniq_l_ctr, bvk_ncells, nbas, batch_size=256):
-    ksh_offsets = []
+def _create_bvk_bas_idx(l_ctr_offsets, uniq_l_ctr, bvk_ncells, nbas):
     ksh_idx = []
-    k0 = k1 = 0
     bvk_bas_offsets = cp.arange(bvk_ncells, dtype=np.int32) * nbas
-    for ksh0, ksh1 in zip(l_ctr_offsets[:-1], l_ctr_offsets[1:]):
-        idx = (bvk_bas_offsets + cp.arange(ksh0, ksh1, dtype=np.int32)[:,None]).ravel()
-        ksh_idx.append(idx)
-        k0, k1 = k1, k1 + len(idx)
-        ksh_offsets.append(cp.arange(k0, k1, batch_size, dtype=np.int32))
-    repeats = [len(x) for x in ksh_offsets]
-    uniq_l_ctr = np.repeat(uniq_l_ctr, repeats, axis=0)
-
-    ksh_offsets.append(np.int32(k1))
-    ksh_offsets = cp.asarray(cp.hstack(ksh_offsets), dtype=np.int32)
+    ksh = cp.arange(l_ctr_offsets[-1], dtype=np.int32)
+    for k0, k1 in zip(l_ctr_offsets[:-1], l_ctr_offsets[1:]):
+        ksh_idx.append((bvk_bas_offsets[:,None] + ksh[k0:k1]).ravel())
     ksh_idx = cp.asarray(cp.hstack(ksh_idx), dtype=np.int32)
-    return ksh_offsets, ksh_idx, uniq_l_ctr
+    return ksh_idx
+
+class _GroupedBatch:
+    batch_range = None
+    shl_range = None
+    ao_range = None
+    sub_batch_offsets = None
+    def __init__(self, **kwargs):
+        self.__dict__.update(kwargs)
+
+def _group_shl_pair_batches(batched_shl_pair_offsets, ao_pair_loc,
+                            ao_pair_batch_size=None, pair_per_block=32):
+    batched_shl_pair_offsets = cp.asnumpy(batched_shl_pair_offsets)
+    if ao_pair_batch_size is None:
+        pair_groups = [
+            _GroupedBatch(
+                batch_range = (0, len(batched_shl_pair_offsets)-1),
+                shl_range = (0, len(ao_pair_loc) - 1),
+                ao_range = (0, ao_pair_loc[-1].get())
+            )
+        ]
+        ao_pair_offsets = np.array(pair_groups[0].ao_range)
+    else:
+        ao_pair_loc_for_group = ao_pair_loc[batched_shl_pair_offsets].get()
+        # Regroup shl_pair batch
+        group_splits = splits_by_blocksize(ao_pair_loc_for_group, ao_pair_batch_size)
+        ao_pair_offsets = ao_pair_loc_for_group[group_splits]
+        pair_groups = [
+            _GroupedBatch(
+                batch_range = (i0, i1),
+                shl_range = (batched_shl_pair_offsets[i0], batched_shl_pair_offsets[i1]),
+                ao_range = (ao_pair_loc_for_group[i0], ao_pair_loc_for_group[i1])
+            ) for i0, i1 in zip(group_splits[:-1], group_splits[1:])
+        ]
+    for group in pair_groups:
+        batch0, batch1 = group.batch_range
+        # Limited by the POOL_SIZE, the PBCsr_int3c2e_latsum23 kernel might not
+        # be able to handle the entire batch in batched_shl_pair_offsets.
+        # Split the batches into sub-batches to ensure load balance and each
+        # sub-batch being fit into POOL_SIZE.
+        sub_batches = [
+            np.arange(batched_shl_pair_offsets[i], batched_shl_pair_offsets[i+1], pair_per_block)
+            for i in range(batch0, batch1)]
+        sub_batches.append(batched_shl_pair_offsets[batch1])
+        group.sub_batch_offsets = asarray(np.hstack(sub_batches, dtype=np.int32))
+    return pair_groups, ao_pair_offsets
+
+def _group_ksh_batches(l_ctr_offsets, uniq_l_ctr, aux_loc, aux_batch_size=None):
+    if aux_batch_size is None:
+        aux_groups = [
+            _GroupedBatch(
+                batch_range = (0, len(l_ctr_offsets)-1),
+                shl_range = (0, len(aux_loc) - 1),
+                ao_range = (0, aux_loc[-1])
+            )
+        ]
+        aux_offsets = np.array(aux_groups[0].ao_range)
+    else:
+        aux_loc_for_group = aux_loc[l_ctr_offsets]
+        # Regroup aux batch into batches. So each kernel launch can handle more aux batches
+        #TODO: group_splits = splits_by_blocksize(aux_loc_for_group, aux_batch_size)
+        group_splits = range(len(l_ctr_offsets))
+        aux_offsets = aux_loc_for_group[group_splits]
+        aux_groups = [
+            _GroupedBatch(
+                batch_range = (i0, i1),
+                shl_range = (l_ctr_offsets[i0], l_ctr_offsets[i1]),
+                ao_range = (aux_loc_for_group[i0], aux_loc_for_group[i1])
+            ) for i0, i1 in zip(group_splits[:-1], group_splits[1:])
+        ]
+    l_ctr_offsets = asarray(l_ctr_offsets, dtype=np.int32)
+    for group in aux_groups:
+        batch0, batch1 = group.batch_range
+        sub_batches = l_ctr_offsets[batch0:batch1+1]
+        group.sub_batch_offsets = sub_batches
+    return aux_groups, aux_offsets
